Decomposedfs propagate diff (#3482)
* rewrite finish upload to get atomic size diff


* decomposedfs: make finish upload atomic


* add lock functions


* allow locking non existing files, fix locking with existing lock


* make returned error recognizable


* more lock fixes


* do not log nil error


* don't overwrite original error when deleting the blob fails


* always release node lock


* keep correct mtimes


* fix adler checksum


* stat before closing


* fix lint ... and proper revision download is not covered by the CS3 api


* fix permissions when downloading grants


* update changelog


* fix locks and revision restore


* assemble permissions on the node when checking a revision


* fix typos


* allow revision download when user has initiate download and list revision permission


* fix reading revision node


* do not forget revision delimiter


* drop old revision


* remove unexpected failures


* update changelog and unexpected passes


* propagate size diff


* fix move, add changelog


* fix size diff calculation for now files


* incorporate reviews


* update tree mock


* fix reva tests


* update tree unit tests


* clarify revisions code


* silence valid return codes


* read size before moving node to trash


* fix s3ng revision download


* clean up revision on error


* use verbose newNode name


* make FlockFile public and more readable


* introduce FileLockSuffix


* use filelock package instead of hardcoded strings


* make linter happy

+++ pkg/storage/utils/decomposedfs/node/node_test.go
@@ -89,10 +89,10 @@ var _ = Describe("Node", func() {
 			n, err := env.Lookup.NodeFromResource(env.Ctx, ref)
 			Expect(err).ToNot(HaveOccurred())
 
-			blobsize := 239485734
+			blobsize := int64(239485734)
 			n.Name = "TestName"
 			n.BlobID = "TestBlobID"
-			n.Blobsize = int64(blobsize)
+			n.Blobsize = blobsize
 
 			err = n.WriteAllNodeMetadata()
 			Expect(err).ToNot(HaveOccurred())
@@ -100,7 +100,7 @@ var _ = Describe("Node", func() {
 			Expect(err).ToNot(HaveOccurred())
 			Expect(n2.Name).To(Equal("TestName"))
 			Expect(n2.BlobID).To(Equal("TestBlobID"))
-			Expect(n2.Blobsize).To(Equal(int64(blobsize)))
+			Expect(n2.Blobsize).To(Equal(blobsize))
 		})
 	})
